src/main/java/com/example/api/dealership/adapter/entrypoint/AuthenticationController.java
package com.example.api.dealership.adapter.entrypoint;

import com.example.api.dealership.adapter.dtos.Response;
import com.example.api.dealership.adapter.dtos.user.UserDtoRequest;
import com.example.api.dealership.adapter.service.security.AuthenticationService;
import io.swagger.v3.oas.annotations.Operation;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@Slf4j
@RestController
@RequestMapping(path = "/v1/dealership")
@RequiredArgsConstructor
public class AuthenticationController {

    private final AuthenticationService authenticationService;

    @Operation(summary = "Authenticate a valid user")
    @PostMapping(path = "/auths",produces = "application/json")
    public ResponseEntity<Response<String>> authenticate(@RequestBody UserDtoRequest userDtoRequest){

        var response = new Response<String>();
        try {
            final var token = authenticationService.authenticate(userDtoRequest);
            response.setData(token);

            return ResponseEntity.ok().body(response);
        }catch (Exception ex){
            response.setErrors("This user is unauthorized");
            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
        }
    }



}
